requirements.postgresql.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
alembic>=1.12.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# PostgreSQL support (replacing PyMySQL)
psycopg2-binary>=2.9.7

# Keep MySQL support for migration period (optional)
# pymysql>=1.1.0

pytest>=7.4.2
httpx>=0.25.0
python-dotenv>=1.0.0
joblib>=1.3.2
pandas>=2.1.0
numpy>=1.26.0
email-validator>=2.0.0
requests>=2.31.0
pyjwt
pytest-cov

# Additional dependencies for migration
PyYAML>=6.0.1
